Beep.AI.MLStudio/Beep.AI.MLStudio/requirements.txt
# Beep ML Studio - Machine Learning Made Simple
# By TheTechIdea
# Core web framework
flask>=3.0.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.1.0
flask-migrate>=4.0.0  # Database migrations with Alembic
flask-jwt-extended>=4.6.0  # JWT authentication

# API and async support
flask-restx>=1.3.0
flask-socketio>=5.3.0

# System utilities
psutil>=5.9.0
python-dotenv>=1.0.0

# UI/Templates
jinja2>=3.1.0
markupsafe>=2.1.0

# WebSocket for real-time updates
python-socketio>=5.10.0
eventlet>=0.33.0

# HTTP client for API communication
requests>=2.31.0
httpx>=0.25.0

# Environment file support
python-dotenv>=1.0.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0

# Visualization
plotly>=5.17.0

# File handling
openpyxl>=3.1.0  # Excel support

# JSON/YAML handling
pyyaml>=6.0.0

# Code execution (for notebooks)
jupyter-client>=8.6.0
ipykernel>=6.25.0

# Note: ML libraries (scikit-learn, tensorflow, pytorch, etc.) 
# will be installed in project-specific virtual environments
# via Beep.Python.Host.Admin

